Django-Backend/requirements.txt
django==3.1.14
djangorestframework==3.13.1
django-extensions==3.1.5
django-timescaledb==0.2.11
psycopg2==2.9.3
channels==3.0.4
paho-mqtt==1.6.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
django-filter==21.1